+++ Android/Components/MainTestingApp/app/src/main/java/com/cajetan/maintestingapp/WebViewActivity.java
@@ -3,11 +3,14 @@ package com.cajetan.maintestingapp;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Base64;
+import android.view.KeyEvent;
+import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
 public class WebViewActivity extends AppCompatActivity {
 
+    // TODO: extend WebView to try to listen for lifecycle events
     WebView mWebView;
 
     @Override
@@ -22,12 +25,26 @@ public class WebViewActivity extends AppCompatActivity {
 
     private void manageWebView() {
         mWebView.setWebViewClient(new WebViewClient());
-        mWebView.loadUrl("http://www.google.com");
+
+        WebSettings webSettings = mWebView.getSettings();
+        webSettings.setJavaScriptEnabled(true);
+
+        mWebView.loadUrl("https://www.youtube.com/watch?v=MTy7ugrSFz4");
 
 //        String unencodedHtml =
 //                "&lt;html&gt;&lt;body&gt;'%23' is the percent code for ‘#‘ &lt;/body&gt;&lt;/html&gt;";
 //        String encodedHtml = Base64.encodeToString(unencodedHtml.getBytes(),
 //                Base64.NO_PADDING);
 //        mWebView.loadData(encodedHtml, "text/html", "base64");
+    }
+
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        if ((keyCode == KeyEvent.KEYCODE_BACK) && mWebView.canGoBack()) {
+            mWebView.goBack();
+            return true;
+        }
+
+        return super.onKeyDown(keyCode, event);
     }
 }